mnist_generation/requirements.txt
pillow
gradio
tensorflow
numpy
opencv-python
keras